% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIME_SimuData.R
\docType{data}
\name{PRIME_SimuData}
\alias{PRIME_SimuData}
\title{prime_SimuData
An Example of Simulated Data for PRIME}
\format{
The dataset prime_SimuData contains n = 200 samples with p = 8 covariates with missing
\describe{
\item{Y}{the response}
\item{X}{the covariates with missing data}
}
}
\usage{
PRIME_SimuData
}
\description{
prime_SimuData
An Example of Simulated Data for PRIME
}
\keyword{datasets}
